Fix path lookup regression on master
Path lookup was not finding the example sources, because
`normalize_path` returns a path that ends with a `/` always.
normalize_path should only be used on the directory component of a
filename.

Also the leading dots should be removed from filenames.

Make check is now finding all the sources again.

Fixes #10

Signed-off-by: Edwin Török <edvin.torok@citrix.com>