4.05/4.06 fix: do not treat all methods as used
`ignore(o)` was treating all methods of `o` as used due to the type of
`o` getting expanded to a `Tobject` with all the methods.
IIUC this is due to same reason as in DeadCode.ml:219, the type is not
present in the source code but gets expanded by the time we reach this
place. If we are dealing with the object itself do not mark its methods
used.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>