[obj] fix FP by fixing the location tracking
During a method call, the identifier in the expression to which
the message is sent used to point to the location on the name of
the inherited class (after the `inherit` keyword), for which we already
created a link with the location of the ancestor.
Now it points to the location of the `inherit` keyword, and we add
an extra link from the keyword to the ancestor.