[obj] expect more precise results on obj/fun_obj_param.ml
The previously expected results were very laxist on the count of uses.
Indeed, passing a factory function as argument would increment the use
counts for all the methods in the returned object.
The new policy is to only increment the use count for methods that are
expected in the corresponding formal parameter, similarly to how
optional arguments are counted as used by type requirements.