[internal] fix --internal ref tracing on dune compiled files
When a file is compiled with dune, the associated builddir is
`/workspace_root` rather than the actual path to the wokspace root.
Consequently, the sourcepath asociated to the current cmti file points
to a non-existent path on the filesystem. This leads `Sys.file_exists`
to fail when looking up for an interface file corresponding to the
current implementation.
`.cmi` files are processed before their corresponding `.cmt` (thanks to
lexical ordering). This means that their sourcepath will be added to the
`abspath` table first. Therefore, finding an interface sourcepath in the
`abspath` is sufficient to confirm that it exists.