[examples] reduce unnecessary recompilation
In 'using_make/', files were systematically compiled on each
invocation of make in 'examples/'. They do not change often and their
compilation produces a lot of compiler warning (which we want to keep
for now).
Now they are only compiled if one of the source file changed. This could
be further improved but does not seem necessary at the time.
As a result, the output of `make examples` is reduced on consecutive
calls and its speed is improved.