[check] fix windows & mac compatibility
The tests in 'examples/using_dune' do not produce the same paths on
windows and unix: on windows we get absolute paths while on unix we get
"parametric" paths, starting with '/workspace_root' which is not
expanded, and followed by the path relative to the dune workspace root.

In 'check/Makefile' the usage of `head -n -1` fails on mac. It is
replaced by `sed '$d'`, and aggregate is updated to return `2` whenever
there is no succeeding test. The issue with `head` lead to a silent
error and a success from `make check` pov because there was no failing
test.