[obj] fix FN on exported immediate objects with interface
In the case of immediate objects (objects without class), the locations
read in the `.cmt` point to the value in the `.ml` rather than in the
`.mli`
Now the link between the value in the `.ml` and the `.mli` is read from
the `cmt_value_dependencies`, and applied before analyzing the implem.
This mirrors the behavior for optional arguments.