[input] read related consecutive cmti files only once
When reading files in a directory, because we sort the content by
filenames, often the`.cmt` is read after its correspnding `.cmi` without
any other `.cmt` or `.cmi` in between. When reading the `.cmi`, the
`.cmt` is also read to gether as many infos as possible. Now, when we
reach the `.cmt`, rather than reading it again, we update the
information stored in the state using pre-loaded data.