[check][make] fix Makefile
Splitting check's output between `.out` and `.err` is a mistake. In case
of FP or FN, they would be in the `.err` and aggregate would not see
them for the unified result. This would lead to an inconsistency between
the summed result indicating failures and unified saying there are none.
It would also lead to aggregate pretending there is no issue (exit code
0) because it relies on the unified result to indicate status.