[examples][dune] add `wrapped_lib` and its uses
`wrapped_lib` is a copy `unwrapped_lib` with a change in the `dune`
fields: `(wrapped true)` (instead of `false`).
The `use_wrapped_lib` is also a copy of `use_unwrapped_lib` with an
extra module indirection to access elements in each `.ml` file.

The `.exp` files have been updated by copying the expected reports for
the `unwrapped_lib` to define expected reports for the `wrapped_lib`.
In the future, new examples should be added in a subdir of `lib`,
symlinked in both `unwrapped_lib` and `wrapped_lib`. Consequently,
the "copy" property of their reports should hold.

The new `.ref` files show that dune's wrapping of module is not an issue
for the tracking of optional arguments and stylistic issues. It is only
an issue for the default reports.