[src][sourcepath] improve sourcepath extraction from cmi
The `sourcepath` was extracted by using the location of the 1st
signature item found in the `cmi_sign`. As exposed by the
`examples/using_make/advanced/incl.ml` test, it would lead to the wrong
compilation unit associated with the module (here `mod.ml` for `Incl`),
resulting in both FP and FN.

Now, the `sourcepath` is the "best candidate" found in all the signature
items. The best candidate is defined in order as:
1. An exact match between the location's sourcefile's compilation unit
   and the cmi's estimated compilation unit.
2. The former is a suffix of the latter.
3. The former is a prefix of the latter
4. Mismatch is still better than no loc.

In case a new candidate would rank the same as the current best one, the
new one is kept.

As a side dev, a new module `Utils` is being defined. The long term goal
is to clear out `DeadCommon` and better organize its code.