[src][input][3/n] store uid's loc for values
Using information available in the `.cmt`, store locations of uids
coming from `include`s in a new state field :
`state.signature.uid_to_loc`. Using `.cmt` and `.cmti`, store locations
of values' (and patterns in value bindings) uids at the same location
for easier lookup.
This is done as soon as the files are loaded, before the analysis.

The new table tries to be semantically equivalent to the usage of
`cmt_value_dependencies` instead of `cmt_declaration_dependencies`
before OCaml 5.3.

This fixes the FP and FN introduced when moving to OCaml 5.3.