[src][out][opt_args] fix missing output in case of colliding locations
If two files are copies of each other, then their internal locations are
equal. In the case of optional arguments, these locations (and labels,
which are also identical in both files) were used as keys in a table.
Thus, optioanl arguments uses from both files were merged and then
reduced by a `sort_uniq` as if there was only one file to begin with.
Consequently, only one of the 2 files would actually see its optional
arguments always/never used reported.
Adding the builddir in the table's key fixes this issue (2 files with
the same name cannot exist withing a unique builddir).

The update of `threshold-3-0.5.ref` show that the missing reports are
now appearing but in the wrong order.