[src][out][opt_args] fix order of reports with colliding locations
When reporting on optional arguments, the different values are sorted by
`abspath` and locations first. When 2 files have the same filename
(found in locations), their `abspath` would be identical too because it
only provides the path of the latest of the 2 files met. Thus, they were
reported together although they belong to different directories.
Using the available builddirs to reconstruct the filepaths instead of
abspath is less error-prone and fixes the ordering issue. Howver, it
incurs the cost of the new strings.