[utils] remove `.pp` extension in `unit`
The compilation unit of a preprocessed file would be `<unit>.pp` when
the compilation unit of its unpreprocessed file would be `<unit>`.
This difference leads to compilation units mismatch when comparing the
`.cmt`'s `sourceunit` and the compilation `unit` of locations
(i.e. linking declarations to their uses), resulting in both FN and FP.
In particular, during `DeadCode.assoc`, the current `.cmt`'s `.ml` can
mistakenly be considered an interface (`is_iface`), although a
corresponding `.mli` exists, because its compilation unit is different
from the `.cmt`'s `sourceunit`.

`Utils.unit` now checks if there is a `.pp` extension in the compilation
unit and removes it to ensure both the original and the preprocessed
files share the same compilation unit.